# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tack API",
    "description": "REST + WebSocket API for Tack, a lightweight, workflow-agnostic project-management tool. This contract is generated from the Rust handlers and domain models; it is the single source of truth for the wire format. All failing responses share the `{ \"error\": { \"status\", \"message\" } }` envelope.",
    "contact": {
      "name": "Tack",
      "email": "info@yielab.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0-beta.6"
  },
  "paths": {
    "/api/attachments/{id}": {
      "get": {
        "tags": [
          "attachments"
        ],
        "summary": "GET /api/attachments/:id",
        "description": "Download an attachment file.",
        "operationId": "download_attachment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Attachment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attachment file bytes",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "description": "Attachment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "attachments"
        ],
        "summary": "DELETE /api/attachments/:id",
        "description": "Delete an attachment file.",
        "operationId": "delete_attachment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Attachment ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Attachment deleted"
          },
          "404": {
            "description": "Attachment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/items/{id}": {
      "get": {
        "tags": [
          "items"
        ],
        "operationId": "get_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Item ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item with roles and dependencies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "items"
        ],
        "operationId": "delete_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Item ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "items"
        ],
        "operationId": "update_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Item ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition / validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/items/{item_id}/attachments": {
      "get": {
        "tags": [
          "attachments"
        ],
        "summary": "GET /api/items/:id/attachments",
        "description": "List all attachments for an item.",
        "operationId": "list_attachments",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "description": "Item ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attachments on the item",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "attachments"
        ],
        "summary": "POST /api/items/:id/attachments",
        "description": "Upload a file attachment to an item.\nAccepts multipart/form-data with a \"file\" field.",
        "operationId": "upload_attachment",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "description": "Item ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Send the binary in a `file` form field (max 50 MB)",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Attachment metadata",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Missing/oversized file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Dependency": {
        "type": "object",
        "required": [
          "id",
          "source_item_id",
          "target_item_id",
          "dependency_type",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "dependency_type": {
            "$ref": "#/components/schemas/DependencyType"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "source_item_id": {
            "type": "string",
            "format": "uuid"
          },
          "target_item_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "DependencyType": {
        "type": "string",
        "enum": [
          "blocks",
          "is_blocked_by",
          "relates_to",
          "duplicates"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable, end-user-facing message.",
            "example": "Item not found"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code, duplicated in the body for convenience.",
            "example": 404,
            "minimum": 0
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "description": "Structured error envelope returned by every failing endpoint — see\n`crate::error::ApiError`. The body is always `{ \"error\": { status, message } }`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
      "EstimateUnit": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "story_points"
            ]
          },
          {
            "type": "string",
            "enum": [
              "hours"
            ]
          },
          {
            "type": "string",
            "enum": [
              "days"
            ]
          },
          {
            "type": "object",
            "required": [
              "custom"
            ],
            "properties": {
              "custom": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Item": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "title",
          "item_type",
          "status",
          "priority",
          "estimate_unit",
          "tags",
          "sort_order",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "assignee": {
            "type": [
              "string",
              "null"
            ]
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "due_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "estimate": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "estimate_unit": {
            "$ref": "#/components/schemas/EstimateUnit"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "item_type": {
            "$ref": "#/components/schemas/ItemType"
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "sort_order": {
            "type": "integer",
            "format": "int32"
          },
          "sprint_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ItemDetail": {
        "type": "object",
        "description": "Detail envelope for `GET /api/items/{id}` — the item plus its assigned roles\nand dependency edges.",
        "required": [
          "item",
          "roles",
          "dependencies"
        ],
        "properties": {
          "dependencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dependency"
            }
          },
          "item": {
            "$ref": "#/components/schemas/Item"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        }
      },
...
      "ItemType": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "epic"
            ]
          },
          {
            "type": "string",
            "enum": [
              "feature"
            ]
          },
          {
            "type": "string",
            "enum": [
              "task"
            ]
          },
          {
            "type": "string",
            "enum": [
              "subtask"
            ]
          },
          {
            "type": "string",
            "enum": [
              "bug"
            ]
          },
          {
            "type": "string",
            "enum": [
              "requirement"
            ]
          },
          {
            "type": "object",
            "required": [
              "custom"
            ],
            "properties": {
              "custom": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Priority": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low",
          "none"
        ]
      },
...
      "Role": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "color",
          "created_at"
        ],
        "properties": {
          "color": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "UpdateItem": {
        "type": "object",
        "properties": {
          "assignee": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "due_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "estimate": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "estimate_unit": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EstimateUnit"
              }
            ]
          },
          "item_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ItemType"
              }
            ]
          },
          "priority": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Priority"
              }
            ]
          },
          "sort_order": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "sprint_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}